fix(ci): portable typescript resolution in hooks gate; rust clippy cast
- scripts/check-hooks.cjs resolved typescript via a hardcoded absolute
  developer path — CI has no such path, so both web and mobile jobs died
  with "Cannot find module 'typescript'". Resolve from cwd first, then
  sibling projects, then plain require.
- web: pin typescript to ~5.9.2 — ^7 installs the native-preview stub
  whose JS API (ScriptTarget/createSourceFile) doesn't exist, which also
  broke the AST hooks check in the web job.
- desktop/nexora-audio: drop u32->u32 cast flagged by clippy -D warnings

+++ desktop/src-tauri/crates/nexora-audio/src/player.rs
@@ -254,7 +254,7 @@ fn run_decode_thread(
                     let (sr, ch) = {
                         let info = dec.info();
                         (
-                            info.sample_rate.max(1) as u32,
+                            info.sample_rate.max(1),
                             info.channels.max(1) as u32,
                         )
                     };
